wip(cachecomponents): extract AttractionCard best-time row to a Client Component
